requirements.txt
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.8
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.16.0
soupsieve==2.6
urllib3==2.2.2
